Implemented rle 2 and added clear screen
- Fixed player death clear
