+++ dungeoneer/src/display/congratulations.asm
@@ -1,6 +1,6 @@
 .local congratulations_map
 intro
-	.byte $ff,$2d,$66,$ff,$1e,$00,$ff,$09,$66,$00,$86,$87,$9e,$9f,$9c,$9d,$8e,$8f,$a4,$a5,$82,$83,$a8,$a9,$aa,$ab,$98,$99,$82,$83,$a8,$a9,$92,$93,$9e,$9f,$9c,$9d,$a6,$a7,$00,$ff,$07,$66,$ff,$0b,$00,$62,$63,$ff,$08,$00,$62,$63,$ff,$0b,$00,$ff,$05,$66,$ff,$08,$00,$38,$ff,$08,$00,$38,$ff,$09,$00,$38,$ff,$08,$00,$66,$66,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$00,$00,$36,$37,$36,$37,$36,$37,$ff,$18,$00,$36,$37,$36,$37,$36,$37,$ff,$04,$00,$6a,$36,$36,$37,$37,$6b,$00,$00,$28,$29,$10,$11,$0a,$0b,$00,$00,$26,$27,$20,$21,$12,$13,$08,$09,$0a,$0b,$24,$25,$00,$00,$6a,$36,$36,$37,$37,$6b,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$05,$00,$22,$23,$2a,$2b,$12,$13,$06,$07,$16,$17,$0a,$0b,$1c,$1d,$26,$27,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$07,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$04,$00
+	.byte $ff,$2d,$66,$ff,$1e,$00,$ff,$09,$66,$00,$86,$87,$9e,$9f,$9c,$9d,$8e,$8f,$a4,$a5,$82,$83,$a8,$a9,$aa,$ab,$98,$99,$82,$83,$a8,$a9,$92,$93,$9e,$9f,$9c,$9d,$a6,$a7,$00,$ff,$07,$66,$ff,$0b,$00,$62,$63,$ff,$08,$00,$62,$63,$ff,$0b,$00,$ff,$05,$66,$ff,$08,$00,$38,$ff,$08,$00,$38,$ff,$09,$00,$38,$ff,$08,$00,$66,$66,$fe,$14,$36,$37,$00,$00,$fe,$03,$36,$37,$ff,$18,$00,$fe,$03,$36,$37,$ff,$04,$00,$6a,$36,$36,$37,$37,$6b,$00,$00,$28,$29,$10,$11,$0a,$0b,$00,$00,$26,$27,$20,$21,$12,$13,$08,$09,$0a,$0b,$24,$25,$00,$00,$6a,$36,$36,$37,$37,$6b,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$05,$00,$22,$23,$2a,$2b,$12,$13,$06,$07,$16,$17,$0a,$0b,$1c,$1d,$26,$27,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$07,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$04,$00
 
 
 .endl
