+++ dungeoneer/src/display/howtoplay.asm
@@ -1,5 +1,5 @@
 .local howtoplay
 map
-	.byte $ff,$30,$00,$38,$00,$10,$11,$1e,$1f,$2e,$2f,$00,$00,$28,$29,$1e,$1f,$00,$00,$20,$21,$18,$19,$02,$03,$32,$33,$00,$38,$ff,$32,$00,$0e,$0f,$0a,$0b,$28,$29,$00,$00,$28,$29,$10,$11,$0a,$0b,$00,$00,$16,$17,$0a,$0b,$32,$33,$ff,$05,$00,$60,$61,$ff,$33,$00,$02,$03,$2c,$2d,$1e,$1f,$12,$13,$08,$09,$00,$00,$08,$09,$0a,$0b,$02,$03,$28,$29,$10,$11,$ff,$05,$00,$64,$64,$00,$65,$65,$00,$c4,$c5,$ff,$2d,$00,$06,$07,$1e,$1f,$18,$19,$18,$19,$0a,$0b,$06,$07,$28,$29,$00,$00,$06,$07,$02,$03,$26,$27,$10,$11,$ff,$03,$00,$68,$69,$00,$3e,$3f,$ff,$30,$00,$0a,$0b,$26,$27,$06,$07,$02,$03,$20,$21,$0a,$0b,$ff,$0f,$00,$58,$59,$ff,$3d,$00,$9a,$9b,$82,$83,$92,$93,$9c,$9d,$00,$00,$9a,$9b,$8a,$8b,$9c,$9d,$aa,$ab,$ff,$08,$00
+	.byte $ff,$30,$00,$38,$00,$10,$11,$1e,$1f,$2e,$2f,$00,$00,$28,$29,$1e,$1f,$00,$00,$20,$21,$18,$19,$02,$03,$32,$33,$00,$38,$ff,$32,$00,$0e,$0f,$0a,$0b,$28,$29,$00,$00,$28,$29,$10,$11,$0a,$0b,$00,$00,$16,$17,$0a,$0b,$32,$33,$ff,$05,$00,$60,$61,$ff,$33,$00,$02,$03,$2c,$2d,$1e,$1f,$12,$13,$08,$09,$00,$00,$08,$09,$0a,$0b,$02,$03,$28,$29,$10,$11,$ff,$05,$00,$64,$64,$00,$65,$65,$00,$c4,$c5,$ff,$2d,$00,$06,$07,$1e,$1f,$fe,$02,$18,$19,$0a,$0b,$06,$07,$28,$29,$00,$00,$06,$07,$02,$03,$26,$27,$10,$11,$ff,$03,$00,$68,$69,$00,$3e,$3f,$ff,$30,$00,$0a,$0b,$26,$27,$06,$07,$02,$03,$20,$21,$0a,$0b,$ff,$0f,$00,$58,$59,$ff,$3d,$00,$9a,$9b,$82,$83,$92,$93,$9c,$9d,$00,$00,$9a,$9b,$8a,$8b,$9c,$9d,$aa,$ab,$ff,$08,$00
 
 .endl
