+++ dungeoneer/src/display/level3.asm
@@ -1,9 +1,9 @@
 .local level3_map
 intro
-	.byte $ff,$04,$66,$ff,$09,$00,$98,$99,$8a,$8b,$ac,$ad,$8a,$8b,$98,$99,$00,$00,$f2,$f3,$ff,$09,$00,$ff,$08,$66,$ff,$05,$00,$26,$27,$20,$21,$12,$13,$16,$17,$0a,$0b,$26,$27,$00,$00,$02,$03,$24,$25,$0a,$0b,$ff,$07,$00,$ff,$08,$66,$ff,$0a,$00,$26,$27,$28,$29,$12,$13,$1c,$1d,$16,$17,$32,$33,$ff,$0a,$00,$ff,$07,$66,$ff,$03,$00,$62,$63,$ff,$18,$00,$62,$63,$ff,$03,$00,$ff,$05,$66,$00,$00,$38,$ff,$08,$00,$38,$ff,$0b,$00,$38,$ff,$08,$00,$38,$ff,$03,$00,$66,$66,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$36,$37,$65,$65,$36,$37,$36,$37,$36,$37,$ff,$18,$65,$36,$37,$36,$37,$36,$37,$65,$65,$00,$00,$6a,$37,$36,$37,$36,$6b,$ff,$18,$00,$6a,$37,$36,$37,$36,$6b,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$1a,$00,$6a,$36,$37,$6b,$ff,$07,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$04,$00
+	.byte $ff,$04,$66,$ff,$09,$00,$98,$99,$8a,$8b,$ac,$ad,$8a,$8b,$98,$99,$00,$00,$f2,$f3,$ff,$09,$00,$ff,$08,$66,$ff,$05,$00,$26,$27,$20,$21,$12,$13,$16,$17,$0a,$0b,$26,$27,$00,$00,$02,$03,$24,$25,$0a,$0b,$ff,$07,$00,$ff,$08,$66,$ff,$0a,$00,$26,$27,$28,$29,$12,$13,$1c,$1d,$16,$17,$32,$33,$ff,$0a,$00,$ff,$07,$66,$ff,$03,$00,$62,$63,$ff,$18,$00,$62,$63,$ff,$03,$00,$ff,$05,$66,$00,$00,$38,$ff,$08,$00,$38,$ff,$0b,$00,$38,$ff,$08,$00,$38,$ff,$03,$00,$66,$66,$fe,$14,$36,$37,$65,$65,$fe,$03,$36,$37,$ff,$18,$65,$fe,$03,$36,$37,$65,$65,$00,$00,$6a,$fe,$02,$37,$36,$6b,$ff,$18,$00,$6a,$fe,$02,$37,$36,$6b,$ff,$05,$00,$6a,$36,$37,$6b,$ff,$1a,$00,$6a,$36,$37,$6b,$ff,$07,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$08,$00,$36,$37,$ff,$1c,$00,$36,$37,$ff,$04,$00
 
 map
-	.byte $10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$08,$0b,$10,$11,$ff,$11,$3f,$05,$04,$05,$04,$05,$0f,$05,$04,$05,$04,$05,$04,$0f,$04,$05,$04,$05,$10,$11,$06,$07,$10,$11,$06,$07,$ff,$0c,$00,$31,$92,$3f,$92,$31,$ff,$0d,$00,$31,$92,$10,$11,$06,$07,$10,$11,$06,$07,$ff,$0d,$00,$31,$3f,$31,$ff,$0f,$00,$31,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$ff,$05,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$ff,$09,$3f,$ff,$05,$00,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$92,$31,$ff,$05,$00,$3f,$ff,$03,$00,$3f,$ff,$0b,$00,$3f,$ff,$05,$00,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$31,$ff,$06,$00,$3f,$ff,$03,$00,$3f,$ff,$0b,$00,$10,$11,$34,$35,$34,$35,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$ff,$09,$3f,$ff,$03,$00,$10,$11,$14,$15,$16,$17,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$ff,$03,$00,$3f,$31,$00,$00,$3f,$ff,$0f,$00,$10,$11,$06,$07,$00,$00,$10,$11,$06,$07,$10,$11,$06,$07,$20,$21,$3f,$00,$00,$31,$3f,$92,$31,$00,$3f,$ff,$0f,$00,$10,$11,$06,$07,$50,$51,$10,$11,$06,$07,$10,$11,$10,$11,$10,$11,$10,$32,$33,$32,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$10,$11,$06,$07,$0a,$0b,$04,$05,$04,$05,$04,$ff,$03,$5d,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$04,$05,$08,$09
+	.byte $fe,$13,$10,$11,$08,$0b,$10,$11,$ff,$11,$3f,$fe,$02,$05,$04,$05,$0f,$fe,$03,$05,$04,$0f,$fe,$02,$04,$05,$10,$11,$06,$07,$10,$11,$06,$07,$ff,$0c,$00,$31,$92,$3f,$92,$31,$ff,$0d,$00,$31,$92,$10,$11,$06,$07,$10,$11,$06,$07,$ff,$0d,$00,$31,$3f,$31,$ff,$0f,$00,$31,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$ff,$05,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$ff,$09,$3f,$ff,$05,$00,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$92,$31,$ff,$05,$00,$3f,$ff,$03,$00,$3f,$ff,$0b,$00,$3f,$ff,$05,$00,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$31,$ff,$06,$00,$3f,$ff,$03,$00,$3f,$ff,$0b,$00,$10,$11,$fe,$02,$34,$35,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$3f,$ff,$03,$00,$ff,$09,$3f,$ff,$03,$00,$10,$11,$14,$15,$16,$17,$10,$11,$06,$07,$10,$11,$06,$07,$00,$00,$3f,$ff,$03,$00,$3f,$31,$00,$00,$3f,$ff,$0f,$00,$10,$11,$06,$07,$00,$00,$10,$11,$06,$07,$10,$11,$06,$07,$20,$21,$3f,$00,$00,$31,$3f,$92,$31,$00,$3f,$ff,$0f,$00,$10,$11,$06,$07,$50,$51,$10,$11,$06,$07,$fe,$03,$10,$11,$10,$32,$33,$32,$fe,$0e,$10,$11,$06,$07,$0a,$0b,$fe,$02,$04,$05,$04,$ff,$03,$5d,$fe,$0d,$04,$05,$04,$05,$08,$09
 
 attrs
 	.byte 1,$b3,$28 ; player pos x/y
@@ -14,7 +14,7 @@ attrs
 	.byte 0
 
 maps1
-	.byte $ff,$05,$92,$11,$10,$ff,$03,$5c,$11,$10,$06,$07,$ff,$1f,$92,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$10,$92,$ff,$07,$31,$ff,$07,$92,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$06,$92,$31,$ff,$0c,$1a,$00,$00,$2a,$2b,$18,$19,$ff,$04,$92,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$06,$92,$1a,$ff,$0c,$30,$36,$00,$2a,$2b,$18,$19,$ff,$03,$92,$31,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$00,$31,$ff,$04,$92,$31,$1a,$ff,$0c,$30,$36,$00,$2a,$2b,$18,$19,$92,$ff,$04,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$00,$00,$ff,$04,$31,$00,$1a,$ff,$04,$30,$ff,$08,$1a,$00,$00,$2a,$2b,$18,$19,$31,$00,$00,$31,$31,$10,$11,$06,$07,$00,$10,$11,$10,$08,$ff,$08,$1a,$ff,$04,$30,$1a,$ff,$0b,$31,$92,$92,$ff,$03,$00,$31,$31,$10,$11,$06,$07,$00,$08,$09,$36,$ff,$0d,$30,$1a,$31,$ff,$0c,$92,$ff,$03,$00,$31,$31,$10,$11,$06,$07,$ff,$03,$00,$36,$ff,$0d,$30,$1a,$31,$ff,$0c,$92,$ff,$03,$00,$31,$31,$10,$11,$10,$11,$10,$11,$10,$11,$08,$ff,$0c,$1a,$31,$ff,$0c,$92,$31,$ff,$03,$00,$31,$31,$08,$09,$04,$ff,$03,$9e,$04,$05,$08,$ff,$0b,$31,$ff,$0e,$92,$31,$ff,$04,$00,$ff,$04,$31,$ff,$19,$92,$ff,$07,$31
+	.byte $ff,$05,$92,$11,$10,$ff,$03,$5c,$11,$10,$06,$07,$ff,$1f,$92,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$10,$92,$ff,$07,$31,$ff,$07,$92,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$06,$92,$31,$ff,$0c,$1a,$00,$00,$2a,$2b,$18,$19,$ff,$04,$92,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$31,$ff,$06,$92,$1a,$ff,$0c,$30,$36,$00,$2a,$2b,$18,$19,$ff,$03,$92,$31,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$00,$31,$ff,$04,$92,$31,$1a,$ff,$0c,$30,$36,$00,$2a,$2b,$18,$19,$92,$ff,$04,$31,$10,$11,$06,$07,$00,$10,$11,$06,$07,$00,$00,$ff,$04,$31,$00,$1a,$ff,$04,$30,$ff,$08,$1a,$00,$00,$2a,$2b,$18,$19,$31,$00,$00,$31,$31,$10,$11,$06,$07,$00,$10,$11,$10,$08,$ff,$08,$1a,$ff,$04,$30,$1a,$ff,$0b,$31,$92,$92,$ff,$03,$00,$31,$31,$10,$11,$06,$07,$00,$08,$09,$36,$ff,$0d,$30,$1a,$31,$ff,$0c,$92,$ff,$03,$00,$31,$31,$10,$11,$06,$07,$ff,$03,$00,$36,$ff,$0d,$30,$1a,$31,$ff,$0c,$92,$ff,$03,$00,$31,$31,$fe,$04,$10,$11,$08,$ff,$0c,$1a,$31,$ff,$0c,$92,$31,$ff,$03,$00,$31,$31,$08,$09,$04,$ff,$03,$9e,$04,$05,$08,$ff,$0b,$31,$ff,$0e,$92,$31,$ff,$04,$00,$ff,$04,$31,$ff,$19,$92,$ff,$07,$31
 
 maps1_attrs
 	.byte 7,8,$ff,$ff,$ff,$ff,$ff,$ff,$ff,$ff ; count, coin state
