+++ dungeoneer/src/display/mainmenu.asm
@@ -1,5 +1,5 @@
 .local mainmenu
 map
-	.byte $ff,$2f,$00,$38,$00,$00,$08,$09,$2a,$2b,$1c,$1d,$0e,$0f,$0a,$0b,$1e,$1f,$1c,$1d,$0a,$0b,$0a,$0b,$24,$25,$00,$00,$38,$ff,$65,$00,$1c,$1d,$0a,$0b,$2e,$2f,$00,$00,$0e,$0f,$02,$03,$1a,$1b,$0a,$0b,$ff,$40,$00,$10,$11,$1e,$1f,$2e,$2f,$00,$00,$28,$29,$1e,$1f,$00,$00,$20,$21,$18,$19,$02,$03,$32,$33,$ff,$3a,$00,$06,$07,$24,$25,$0a,$0b,$08,$09,$12,$13,$28,$29,$26,$27,$ff,$84,$00
+	.byte $ff,$2f,$00,$38,$00,$00,$08,$09,$2a,$2b,$1c,$1d,$0e,$0f,$0a,$0b,$1e,$1f,$1c,$1d,$fe,$02,$0a,$0b,$24,$25,$00,$00,$38,$ff,$65,$00,$1c,$1d,$0a,$0b,$2e,$2f,$00,$00,$0e,$0f,$02,$03,$1a,$1b,$0a,$0b,$ff,$40,$00,$10,$11,$1e,$1f,$2e,$2f,$00,$00,$28,$29,$1e,$1f,$00,$00,$20,$21,$18,$19,$02,$03,$32,$33,$ff,$3a,$00,$06,$07,$24,$25,$0a,$0b,$08,$09,$12,$13,$28,$29,$26,$27,$ff,$84,$00
 
 .endl
