+++ dungeoneer/src/interrupt.asm
@@ -72,7 +72,6 @@ vvblkd_interrupt
     render_intro()
     render_gameover()
     render_player_death()
-
     render_enemy()
 done
     ; restore stack
